fixed allowed clients ip check logic. added LOG info to trace some operations.

+++ src/main/java/org/appxi/api/pieces/controller/ServiceController.java
@@ -1,6 +1,8 @@
 package org.appxi.api.pieces.controller;
 
 import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.appxi.api.pieces.model.Piece;
 import org.appxi.api.pieces.repo.db.PiecePersistRepository;
 import org.appxi.api.pieces.repo.solr.PieceSolrRepository;
@@ -15,6 +17,8 @@ import java.util.Set;
 @RestController
 @RequestMapping("/api")
 class ServiceController {
+    private static final Log logger = LogFactory.getLog(ServiceController.class);
+
     private final PiecePersistRepository persistRepository;
     private final PieceSolrRepository solrRepository;
     private final Set<String> allowedClients;
@@ -28,8 +32,10 @@ class ServiceController {
     @GetMapping("/piece/{id}")
     public Piece detail(@PathVariable("id") String id,
                         HttpServletResponse resp) throws IOException {
-        if (StringUtils.isBlank(id))
+        if (StringUtils.isBlank(id)) {
             resp.sendError(400);
+            return null;
+        }
         Piece result = persistRepository.findById(id).orElse(null);
         if (null == result)
             resp.sendError(404);
@@ -39,9 +45,12 @@ class ServiceController {
     @PostMapping("/piece")
     public String create(@RequestBody List<Piece> entries,
                          HttpServletRequest request, HttpServletResponse resp) throws IOException {
-        verifyAllowedClient(request, resp);
-        if (null == entries || entries.isEmpty())
+        if (isNotAllowedClient(request, resp))
+            return null;
+        if (null == entries || entries.isEmpty()) {
             resp.sendError(400);
+            return null;
+        }
         Iterable<Piece> savedList = persistRepository.saveAll(entries);
         solrRepository.saveAll(savedList);
         return "created";
@@ -50,9 +59,12 @@ class ServiceController {
     @PutMapping("/piece")
     public String update(@RequestBody List<Piece> entries,
                          HttpServletRequest request, HttpServletResponse resp) throws IOException {
-        verifyAllowedClient(request, resp);
-        if (null == entries || entries.isEmpty())
+        if (isNotAllowedClient(request, resp))
+            return null;
+        if (null == entries || entries.isEmpty()) {
             resp.sendError(400);
+            return null;
+        }
         Iterable<Piece> savedList = persistRepository.saveAll(entries);
         solrRepository.saveAll(savedList);
         return "updated";
@@ -61,9 +73,12 @@ class ServiceController {
     @DeleteMapping("/piece/{id}")
     public String delete(@PathVariable("id") String id,
                          HttpServletRequest request, HttpServletResponse resp) throws IOException {
-        verifyAllowedClient(request, resp);
-        if (StringUtils.isBlank(id))
+        if (isNotAllowedClient(request, resp))
+            return null;
+        if (StringUtils.isBlank(id)) {
             resp.sendError(400);
+            return null;
+        }
         String result = null;
         try {
             persistRepository.deleteById(id);
@@ -82,10 +97,32 @@ class ServiceController {
         return result;
     }
 
-    private void verifyAllowedClient(HttpServletRequest req, HttpServletResponse resp) throws IOException {
+
+    @PostMapping("/service/ip/{ip}")
+    public String serviceIpAdd(@PathVariable("ip") String ip,
+                               HttpServletRequest request, HttpServletResponse resp) throws IOException {
+        if (isNotAllowedClient(request, resp))
+            return null;
+        allowedClients.add(ip);
+        return "added";
+    }
+
+    @DeleteMapping("/service/ip/{ip}")
+    public String serviceIpRemove(@PathVariable("ip") String ip,
+                                  HttpServletRequest request, HttpServletResponse resp) throws IOException {
+        if (isNotAllowedClient(request, resp))
+            return null;
+        allowedClients.remove(ip);
+        return "removed";
+    }
+
+    private boolean isNotAllowedClient(HttpServletRequest req, HttpServletResponse resp) throws IOException {
         String remoteIp = req.getRemoteAddr();
         boolean allowed = allowedClients.contains(remoteIp);
-        if (!allowed)
+        if (!allowed) {
+            logger.warn("Denied: " + remoteIp);
             resp.sendError(403);
+        }
+        return !allowed;
     }
 }
